feat: add REST endpoint

+++ flyway/sql/migrations/V201121.1712__create_tables.sql
@@ -1,7 +1,7 @@
 -- Address Table
 CREATE TABLE flyway_demo_db.address
 (
-    id SERIAL PRIMARY KEY,
+    id NUMERIC(19,0) PRIMARY KEY,
     address_1 VARCHAR(250),
     city VARCHAR(50),
     state VARCHAR(50),
